Squash nested loop

+++ src/routing/polygon_traversal.jl
@@ -378,7 +378,7 @@ function is_visible(
     current_point::Point{2,Float64},
     final_point::Point{2,Float64},
     exclusions::DataFrame,
-    current_exclusions_idx::Vector{Int}=[0]
+    current_exclusions_idx::Vector{Int64}=[0]
 )::Bool
     return is_visible(current_point, final_point, exclusions.geometry, current_exclusions_idx)
 end
